Add Full CRUD

+++ src/main/java/com/example/aucaregistration/repository/AcademicUnitRepository.java
@@ -3,5 +3,5 @@ package com.example.aucaregistration.repository;
 import com.example.aucaregistration.domain.AcademicUnit;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AcademicUnitRepository extends JpaRepository<AcademicUnit, Integer> {
+public interface AcademicUnitRepository extends JpaRepository<AcademicUnit, Long> {
 }
