+++ src/main/java/com/example/aucaregistration/repository/CourseDefinitionRepository.java
@@ -3,5 +3,5 @@ package com.example.aucaregistration.repository;
 import com.example.aucaregistration.domain.CourseDefinition;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CourseDefinitionRepository extends JpaRepository<CourseDefinition, Integer> {
+public interface CourseDefinitionRepository extends JpaRepository<CourseDefinition, Long> {
 }
