+++ src/main/java/com/example/aucaregistration/repository/SemesterRepository.java
@@ -3,5 +3,5 @@ package com.example.aucaregistration.repository;
 import com.example.aucaregistration.domain.Semester;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SemesterRepository extends JpaRepository<Semester, Integer> {
+public interface SemesterRepository extends JpaRepository<Semester, Long> {
 }
